MGN-pytorch-master/requirements.txt
torchvision
matplotlib
argparse
sklearn
pytorch
pillow
numpy
scipy
tqdm
